Fix query cost estimation for ondisk storage. Add type parameterized tests for query cost estimation.

+++ src/storage/v2/disk/rocksdb_storage.hpp
@@ -45,7 +45,7 @@ struct RocksDBStorage {
   uint64_t ApproximateVertexCount() const {
     uint64_t estimate_num_keys = 0;
     db_->GetIntProperty(vertex_chandle, "rocksdb.estimate-num-keys", &estimate_num_keys);
-    return static_cast<int64_t>(estimate_num_keys);
+    return estimate_num_keys;
   }
 };
